refactored show_manager for modularity
