+++ backend/shows/__init__.py
@@ -1,0 +1,12 @@
+from .show_ingest import fetch_and_store_show, fast_fetch_and_store_show
+from .show_refresh import process_missing_refresh, process_show_refresh, process_metadata_refresh
+from .show_enrich import _enrichment_in_progress
+
+__all__ = [
+    'fetch_and_store_show',
+    'fast_fetch_and_store_show',
+    'process_missing_refresh',
+    'process_show_refresh',
+    'process_metadata_refresh',
+    '_enrichment_in_progress'
+]
